+++ src/components/ReleateNotes/SavedSchedulesDiff.tsx
@@ -1,22 +1,11 @@
-import { Clock, User } from "lucide-react";
 import { type ComponentProps, Fragment, useMemo } from "react";
-import LecLabComponent from "src/components/LecLab";
-import TeacherStats from "src/components/TeacherStats";
 import { useSavedSchedule } from "src/hooks";
-import { Iter } from "src/lib/iter";
 import isValidAdditionToSchedule from "src/lib/schedule/isValidAdditionToSchedule";
 import { useSectionStore } from "src/lib/store/section";
-import { capitalize, cn } from "src/lib/utils";
-import type { DayTime, LecLab, Section } from "src/types/generated";
-
-const red = "red-300";
-const green = "green-300";
-
-const redBg = `bg-${red}`;
-const greenBg = `bg-${green}`;
-
-const redText = `text-${red}`;
-const greenText = `text-${green}`;
+import { cn } from "src/lib/utils";
+import type { Section } from "src/types/generated";
+import { greenBg, redBg, redText } from "./colors";
+import { Diff, LeclabsDiff } from "./diffs";
 
 type ModifiedSection = {
   oldSection: Section;
@@ -257,168 +246,6 @@ function SectionTimes({
   });
 }
 
-function LeclabsDiff({
-  sectionId,
-  oldLeclabs,
-  newLecLabs,
-}: {
-  sectionId: string;
-  oldLeclabs: LecLab[];
-  newLecLabs: LecLab[];
-}) {
-  const minLen = Math.min(oldLeclabs.length, newLecLabs.length);
-
-  return (
-    <>
-      {Iter.from(oldLeclabs)
-        .zip(Iter.from(newLecLabs))
-        .map(([oldLeclab, newLeclab], i) => (
-          <LecLabDiff
-            key={`${sectionId}-leclabdiff-${JSON.stringify(oldLeclab.dayTimes)}-${JSON.stringify(newLeclab.dayTimes)}-${i.toString()}`}
-            oldLeclab={oldLeclab}
-            newLeclab={newLeclab}
-          />
-        ))}
-      {Iter.from(oldLeclabs)
-        .skip(minLen)
-        .map((leclab, i) => (
-          <LecLabComponent
-            key={`${sectionId}-extraoldleclab-${JSON.stringify(leclab.dayTimes)}-${i.toString()}`}
-            leclab={leclab}
-            className={cn(redText, "bg-red-900")}
-          />
-        ))}
-      {Iter.from(newLecLabs)
-        .skip(minLen)
-        .map((leclab, i) => (
-          <LecLabComponent
-            key={`${sectionId}-extranewleclab-${JSON.stringify(leclab.dayTimes)}-${i.toString()}`}
-            leclab={leclab}
-            className={cn(greenText, "bg-green-900")}
-          />
-        ))}
-    </>
-  );
-}
-
-function LecLabDiff({
-  oldLeclab,
-  newLeclab,
-}: {
-  oldLeclab: LecLab;
-  newLeclab: LecLab;
-}) {
-  const oldLeclabType = capitalize(oldLeclab.type ?? "lecture");
-  const newLeclabType = capitalize(newLeclab.type ?? "lecture");
-
-  return (
-    <div className={cn("bg-secondary rounded-md p-2")}>
-      <h4 className="italic">
-        {oldLeclabType === newLeclabType ? (
-          oldLeclabType
-        ) : (
-          <Diff oldStr={oldLeclabType} newStr={newLeclabType} />
-        )}
-      </h4>
-
-      <div className="relative flex items-center gap-2">
-        <User className="h-4 opacity-50" />
-        {oldLeclab.prof === newLeclab.prof ? (
-          <>
-            {oldLeclab.prof}
-            <TeacherStats leclab={oldLeclab} />
-          </>
-        ) : (
-          <div className="flex w-full flex-col">
-            <div className={cn("flex gap-1", redText)}>
-              {oldLeclab.prof}
-              <TeacherStats leclab={oldLeclab} />
-            </div>
-            <div className={cn("flex gap-1", greenText)}>
-              {newLeclab.prof}
-              <TeacherStats leclab={newLeclab} />
-            </div>
-          </div>
-        )}
-      </div>
-
-      <DaytimesDiff
-        title={oldLeclab.title}
-        oldDaytimes={oldLeclab.dayTimes}
-        newDaytimes={newLeclab.dayTimes}
-      />
-    </div>
-  );
-}
-
-function DaytimesDiff({
-  title,
-  oldDaytimes,
-  newDaytimes,
-}: {
-  title: string;
-  oldDaytimes: DayTime[];
-  newDaytimes: DayTime[];
-}) {
-  const minLen = Math.min(oldDaytimes.length, newDaytimes.length);
-
-  return (
-    <>
-      {Iter.from(oldDaytimes)
-        .zip(Iter.from(newDaytimes))
-        .map(([oldDaytime, newDaytime]) => (
-          <div
-            className="flex items-center gap-2"
-            key={
-              oldDaytime.day +
-              oldDaytime.startTimeHhmm +
-              oldDaytime.endTimeHhmm +
-              newDaytime.day +
-              newDaytime.startTimeHhmm +
-              newDaytime.endTimeHhmm +
-              title
-            }
-          >
-            <Clock className="h-4 opacity-50" />
-            <Diff oldStr={oldDaytime.day} newStr={newDaytime.day} />{" "}
-            <Diff
-              oldStr={`${oldDaytime.startTimeHhmm}-${oldDaytime.endTimeHhmm}`}
-              newStr={`${newDaytime.startTimeHhmm}-${newDaytime.endTimeHhmm}`}
-            />
-          </div>
-        ))}
-
-      {Iter.from(oldDaytimes)
-        .skip(minLen)
-        .map((dayTime) => (
-          <div
-            className={cn("flex items-center gap-2", redText)}
-            key={
-              dayTime.day + dayTime.startTimeHhmm + dayTime.endTimeHhmm + title
-            }
-          >
-            <Clock className="h-4 opacity-50" />
-            {dayTime.day} {dayTime.startTimeHhmm}-{dayTime.endTimeHhmm}
-          </div>
-        ))}
-
-      {Iter.from(newDaytimes)
-        .skip(minLen)
-        .map((dayTime) => (
-          <div
-            className={cn("flex items-center gap-2", greenText)}
-            key={
-              dayTime.day + dayTime.startTimeHhmm + dayTime.endTimeHhmm + title
-            }
-          >
-            <Clock className="h-4 opacity-50" />
-            {dayTime.day} {dayTime.startTimeHhmm}-{dayTime.endTimeHhmm}
-          </div>
-        ))}
-    </>
-  );
-}
-
 function OverlappingDiv({
   children,
   className,
@@ -442,64 +269,6 @@ function OverlappingDiv({
         )}
       />
       {children}
-    </div>
-  );
-}
-
-function Diff({ oldStr, newStr }: { oldStr: string; newStr: string }) {
-  return oldStr === newStr ? (
-    oldStr
-  ) : (
-    <>
-      <span className={cn(redText)}>{oldStr}</span>
-      {" → "}
-      <span className={cn(greenText)}>{newStr}</span>
-    </>
-  );
-}
-
-type ChangedSectionPreviewCardProps = {
-  oldSection: Section;
-  newSection: Section;
-  className?: string;
-};
-export function ChangedSectionPreviewCard({
-  oldSection,
-  newSection,
-  className,
-}: ChangedSectionPreviewCardProps) {
-  return (
-    <div
-      className={cn(
-        "bg-secondary/50 flex flex-col gap-2 rounded-xl p-2",
-        className,
-      )}
-    >
-      <div>
-        <h2 className="font-light">
-          <Diff oldStr={oldSection.course} newStr={newSection.course} />
-          {": "}
-          <Diff oldStr={oldSection.domain} newStr={newSection.domain} />{" "}
-          <Diff oldStr={oldSection.code} newStr={newSection.code} />
-        </h2>
-
-        <h1 className="font-heading text-base font-bold md:text-xl">
-          <Diff oldStr={oldSection.section} newStr={newSection.section} />{" "}
-          <Diff oldStr={oldSection.title} newStr={newSection.title} />
-        </h1>
-      </div>
-
-      <LeclabsDiff
-        oldLeclabs={oldSection.leclabs}
-        newLecLabs={newSection.leclabs}
-        sectionId={oldSection.id}
-      />
-
-      {oldSection.more !== "" || newSection.more !== "" ? (
-        <p className="opacity-70">
-          <Diff oldStr={oldSection.more} newStr={newSection.more} />
-        </p>
-      ) : null}
     </div>
   );
 }
