+++ src/components/ReleateNotes/colors.ts
@@ -1,0 +1,4 @@
+export const redBg = `bg-red-300`;
+export const greenBg = `bg-green-300`;
+export const redText = `text-red-300`;
+export const greenText = `text-green-300`;
