+++ src/components/ReleateNotes/index.tsx
@@ -1,0 +1,2 @@
+export * from "./ReleaseNotes";
+export * from "./SavedSchedulesDiff";
